Center the canvas on the screen and set it's width
A canvas element actually has two sizes that need to be considered: the
size of the DOM element, which is how much screen space your canvas will
take up, and the size of the drawing surface, which is how many points
you can effectively draw to on the page. These two values are
canvas.style.{width,height} and canvas.{width,height}, respectively, if
you are setting them using JavaScript. If you need to set these
statically, for the first, use the width and height properties on the
DOM element. For the second, use the width and height properties in a
CSS rule on of the appropriate selector.

If these numbers are different, then the browser will scale the drawing
canvas to fit inside of the screen size, potentially distorting any
drawings that you create. If canvas.{width,height} is/are set and the
other corresponding style isn't set, then the browser will set it for
you. However, this does not work the other way around: if
canvas.style.{width,height} is set and the other corresponding property
isn't, the default values of 300px for width and 150px for height will
be used.